backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
apscheduler==3.10.4
weasyprint==63.1
jinja2==3.1.5
aiofiles==24.1.0
pillow==11.1.0
